+++ CHANGELOG.md
@@ -30,6 +30,7 @@ the abandoned `oomphinc/composer-installers-extender` project:
 
 - [Update Drupal core to 11.4 #1096](https://github.com/farmOS/farmOS/pull/1096)
 - [Require plan records to reference a plan #1094](https://github.com/farmOS/farmOS/pull/1094)
+- [Allow referencing multiple genetic fathers in birth quick form #1101](https://github.com/farmOS/farmOS/pull/1101)
 
 ### Deprecated